+++ tdmclient/tools/__init__.py
@@ -1,2 +1,4 @@
 """Tools to be launched from a terminal or from outside Python.
 """
+
+from . import (gui, list, repl, run, server, tdmdiscovery, transpile, watch)